styled navbar, renamed folders
