+++ frontend/src/components/TripList/NewTripModal/index.jsx
@@ -1,0 +1,3 @@
+import React from "react";
+
+// TODO implement functionality to add new trips
