+++ frontend/src/components/TripList/Trip/Searchbar/index.jsx
