+++ frontend/src/components/TripList/Trip/index.jsx
