+++ frontend/src/components/TripList/index.jsx
@@ -1,10 +1,11 @@
 import React from "react";
-import Trip from "./SearchBar";
+import Trip from "./Trip";
 
 function TripList() {
     return (
         <div>
             <Trip />
+            {/* TODO show a list of trips */}
         </div>
     );
 }
